Moves RegistryValueKind and RegistryView to shared and Improves diff in RegistryKey
