+++ src/System.Private.CoreLib/shared/Microsoft/Win32/RegistryValueKind.cs
@@ -11,13 +11,13 @@ namespace Microsoft.Win32
 #endif
     enum RegistryValueKind
     {
-        String = Interop.mincore.RegistryValues.REG_SZ,
-        ExpandString = Interop.mincore.RegistryValues.REG_EXPAND_SZ,
-        Binary = Interop.mincore.RegistryValues.REG_BINARY,
-        DWord = Interop.mincore.RegistryValues.REG_DWORD,
-        MultiString = Interop.mincore.RegistryValues.REG_MULTI_SZ,
-        QWord = Interop.mincore.RegistryValues.REG_QWORD,
+        String = Interop.Kernel32.RegistryValues.REG_SZ,
+        ExpandString = Interop.Kernel32.RegistryValues.REG_EXPAND_SZ,
+        Binary = Interop.Kernel32.RegistryValues.REG_BINARY,
+        DWord = Interop.Kernel32.RegistryValues.REG_DWORD,
+        MultiString = Interop.Kernel32.RegistryValues.REG_MULTI_SZ,
+        QWord = Interop.Kernel32.RegistryValues.REG_QWORD,
         Unknown = 0,                          // REG_NONE is defined as zero but BCL
         None = unchecked((int)0xFFFFFFFF), //  mistakenly overrode this value.  
-    }   // Now instead of using Interop.mincore.RegistryValues.REG_NONE we use "-1".
+    }   // Now instead of using Interop.Kernel32.RegistryValues.REG_NONE we use "-1".
 }
