+++ src/System.Private.CoreLib/shared/Microsoft/Win32/RegistryView.cs
@@ -14,7 +14,7 @@ namespace Microsoft.Win32
     enum RegistryView
     {
         Default = 0, // 0x0000 operate on the default registry view
-        Registry64 = Interop.mincore.RegistryView.KEY_WOW64_64KEY, // 0x0100 operate on the 64-bit registry view
-        Registry32 = Interop.mincore.RegistryView.KEY_WOW64_32KEY, // 0x0200 operate on the 32-bit registry view
+        Registry64 = Interop.Kernel32.RegistryView.KEY_WOW64_64KEY, // 0x0100 operate on the 64-bit registry view
+        Registry32 = Interop.Kernel32.RegistryView.KEY_WOW64_32KEY, // 0x0200 operate on the 32-bit registry view
     };
 }
